Step 5 of Lesson 1 Homework
Homework prompt: (Practical) Create a simple system.out app. Print info about the user that should be passed to the main function as a parameter. Compile and run the program from the command line using javac and java from JDK.

Created two examples. One that allows String arguments to be passed into the main function to create a Student Profile, and another that allows int arguments to be passed into the main function to calculate the area of a rectangle.